backend/requirements.txt
Flask==2.2.3
Flask_Cors==4.0.0
flask_mongoengine==1.0.0
Flask_RESTful==0.3.10
langchain==0.0.292
mongoengine==0.27.0
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.2.3
gunicorn>=20.0
sentence-transformers==2.2.2
langchain_google_genai==0.0.6
cloudinary==1.38.0